// ui/src/sdk/Resources.ts
export enum Resources {


    ModuleTruck = 'truck',

    TTFKenpixel = 'kenpixel',
    Area = "area",
    Shift = "shift",
    Road = "road", // 道路箭头
    Goods = "goods", // 货物
    Goods0 = "goods_0", // 货物
    Goods1 = "goods_1", // 货物
    Goods2 = "goods_2", // 货物
    Cell = "cell", // 槽位
    Floor = "floor", // 底板
    Rack = "rack", // 支架
    SkyScene = "skyScene", // 天空背景
    Sky = "sky" // 天空

}